libft == norme

+++ ft_calloc.c
@@ -1,16 +1,16 @@
 #include <stdlib.h>
 
-void	*ft_calloc(int nmemb, int size)
+void	*ft_calloc(size_t nmemb, size_t size)
 {
-	int i;
-	char *mem;
-	
+	size_t	i;
+	char	*mem;
+
 	mem = malloc(nmemb * size);
 	i = 0;
-	while(i < nmemb)
+	while (i < nmemb)
 	{
 		mem[i] = 0;
 		i++;
 	}
-	return(mem);
+	return (mem);
 }
